+++ src/components/PageFooter/PageFooter.tsx
@@ -1,5 +1,7 @@
 import { FC } from 'react';
-import { Footer } from '@taskany/bricks';
+import { Footer, Link } from '@taskany/bricks';
+import { FooterItem } from '@taskany/bricks/components/Footer';
+import { gray9 } from '@taskany/colors';
 
 import { tr } from './PageFooter.i18n';
 
@@ -11,5 +13,14 @@ export const PageFooter: FC = () => {
         { title: tr('API'), url: '/api' },
         { title: tr('About'), url: '/about' },
     ];
-    return <Footer menuItems={menuItems} />;
+
+    return (
+        <Footer>
+            {menuItems.map(({ title, url }) => (
+                <Link key={url} href={url} inline>
+                    <FooterItem color={gray9}>{title}</FooterItem>
+                </Link>
+            ))}
+        </Footer>
+    );
 };